Added randomSong Method
There is an error in the stop() method which is supposed to stop the
music play. This is neither my mistake in code nor anything I did to the
MusicPlayer. When I wrote the same exact code into the BlueJ Projects
MusicPlayer v5 my code worked just as well and there was no issue with
the stop() method.